Renamed time-based component and folder
