+++ src/components/Session/Form/Startup.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Countdown} from '../../Clock/Countdown'
+import {Countdown} from '../../Time/Countdown'
 import {WorkoutExerciseItem} from '../../Workout/View/ExerciseItem'
 
 export const SessionStartup = ({nextExercise, finishAt, onSessionStartupCompleted}) => (
