+++ src/components/Session/Form/__tests__/Exercise.spec.js
@@ -4,8 +4,8 @@ import sinon from 'sinon'
 import {expect} from 'chai'
 import {shallow} from 'enzyme'
 import {SessionExercise} from '../Exercise'
-import {Timer} from '../../../Clock/Timer'
-import {Countdown} from '../../../Clock/Countdown'
+import {Clock} from '../../../Time/Clock'
+import {Countdown} from '../../../Time/Countdown'
 import {WorkoutExerciseItem} from '../../../Workout/View/ExerciseItem'
 import {UNITS} from '../../../../api/unit'
 const moment = require('moment')
@@ -28,7 +28,7 @@ describe('<SessionExercise/>', () => {
     const wrapper = shallow(<SessionExercise {...props}/>)
     expect(wrapper.find(WorkoutExerciseItem)).to.have.lengthOf(1)
     expect(wrapper.find(Countdown)).to.have.lengthOf(0)
-    expect(wrapper.find(Timer)).to.have.lengthOf(1)
+    expect(wrapper.find(Clock)).to.have.lengthOf(1)
     expect(wrapper.find('button').text()).to.be.equal('Done')
   })
 
@@ -55,8 +55,8 @@ describe('<SessionExercise/>', () => {
     expect(wrapper.find(Countdown).props().onCountdownCompleted).to.be.equal(props.onExerciseCompleted)
     expect(wrapper.find(WorkoutExerciseItem)).to.have.lengthOf(1)
 
-    // Verify <Timer/> and other components are not rendered
-    expect(wrapper.find(Timer)).to.have.lengthOf(0)
+    // Verify <Clock/> and other components are not rendered
+    expect(wrapper.find(Clock)).to.have.lengthOf(0)
     expect(wrapper.find('button')).to.have.lengthOf(0)
   })
 })
