+++ src/components/Session/View/__tests__/View.spec.js
@@ -3,7 +3,7 @@ import {Factory} from 'rosie'
 import {shallow} from 'enzyme'
 import {expect} from 'chai'
 import {SessionView} from '../View'
-import {Duration} from '../../../Clock/Duration'
+import {Duration} from '../../../Time/Duration'
 import {RoundsCompleted} from '../RoundsCompleted'
 import {WorkoutSetup} from '../../../Workout/View/Setup'
 import {WorkoutExerciseList} from '../../../Workout/View/ExerciseList'
