+++ src/components/Time/Duration.js
