+++ src/components/Time/__tests__/Duration.spec.js
